Allow broken sky blocks to stack with their respective types
The previous behavior was that sky blocks would drop an ItemStack with
the metadata in world. e.g. if a normal sky block was active in the
world (metadata 2) and was broken, the resulting stack would have
metadata 2 and not be able to stack with normal, crafted sky blocks
(metadata 0) and have to be replaced in the world, and forced to update
to restore them to metadata 0. Now, active and inactive normal sky
blocks always drop metadata 0, and active and inactive inverted sky
blocks always drop metadata 1