Change elevators so players don't end up in walls
This changes elevators so that if the player was to end up (partially) inside a wall after using the elevator, they instead end up on center. If there's no wall at target location, it should behave like before. Could potentially be improved to only move the player away from the wall on one axis when possible.